+++ packages/nyxojs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # nyxo.js
 
+## 0.3.0
+
+### Minor Changes
+
+- update pinned messages handling; add pagination support and enhance role color configurations
+
+### Patch Changes
+
+- Updated dependencies
+  - @nyxojs/builders@0.3.0
+  - @nyxojs/gateway@0.3.0
+  - @nyxojs/store@0.3.0
+  - @nyxojs/core@0.3.0
+  - @nyxojs/rest@0.3.0
+
 ## 0.2.6
 
 ### Patch Changes