+++ packages/zstd/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nyxojs/zstd
 
+## 0.2.0
+
+### Minor Changes
+
+- update pinned messages handling; add pagination support and enhance role color configurations
+
 ## 0.1.0
 
 ### Minor Changes